Start testing queue methods.
