+++ data_structures/queue/queue.py
@@ -1,18 +1,22 @@
+from .node import Node
+
+
 class Queue:
     def __init__(self, it=()):
         """
         Initialize new list with optional iterable.
         """
         self.head = None
+        self._size = 0
 
         for value in it:
             self.push(value)
 
-    def peek(self):
-        pass
+    def enqueue(self, value):
+        self.head = Node(value, self.head)
+        self._size += 1
 
-    def pop(self):
-        pass
-
-    def push(self, value):
-        pass
+    def dequeue(self):
+        if not self.head:
+            raise IndexError('')
+        self._size -= 1
